✅ (GURB): Fix create_initial_invoice test
