+++ som_gurb/data/som_gurb_data.xml
@@ -20,7 +20,7 @@
             <field name="categ_id" ref="categ_product_gurb"/>
             <field name="type">service</field>
             <field name="procure_method">make_to_stock</field>
-            <field name="standard_price">10</field>
+            <field name="list_price">10</field>
             <field name="supply_method">buy</field>
         </record>
         <record model="product.product" id="product_owner_gurb">
@@ -28,7 +28,7 @@
             <field name="categ_id" ref="categ_product_gurb"/>
             <field name="type">service</field>
             <field name="procure_method">make_to_stock</field>
-            <field name="standard_price">0</field>
+            <field name="list_price">0</field>
             <field name="supply_method">buy</field>
         </record>
         <record model="product.product" id="initial_quota_gurb">
