+++ som_gurb/tests/__init__.py
@@ -1,3 +1,4 @@
 from tests_gurb_switching import *
 from tests_gurb_services import *
 from tests_gurb_cups import *
+from tests_gurb_initial_quota import *
